+++ ProyectosSDL/HolaSDL/StickyAward.h
@@ -4,6 +4,6 @@ class StickyAward : public Award
 {
 public:
   StickyAward(){};
-  StickyAward(float32 x, float32 y, float32 width, float32 height, float32 speed, float32 framerate, Texture *texture);
+  StickyAward(float32 x, float32 y, float32 width, float32 height, float32 speed, uint framerate, Texture *texture);
   virtual void contact();
 };
